# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SubTrack – Subscription Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SubscriptionBillingCycle = typeof SubscriptionBillingCycle[keyof typeof SubscriptionBillingCycle];


export const SubscriptionBillingCycle = {
  weekly: 'weekly',
  monthly: 'monthly',
  quarterly: 'quarterly',
  semi_annual: 'semi_annual',
  yearly: 'yearly',
} as const;

export interface Subscription {
  id: number;
  userId: string;
  name: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  price: number;
  currency: string;
  billingCycle: SubscriptionBillingCycle;
  /** ISO date string YYYY-MM-DD */
  renewalDate: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  notes?: string | null;
  isActive: boolean;
  isArchived: boolean;
  /** Monthly cost equivalent in the original currency */
  monthlyEquivalent: number;
  /** Annual cost equivalent in the original currency */
  annualEquivalent: number;
  /**
     * Days until next renewal (null if archived)
     * @nullable
     */
  daysUntilRenewal?: number | null;
  createdAt: string;
  updatedAt: string;
}

export type SubscriptionInputBillingCycle = typeof SubscriptionInputBillingCycle[keyof typeof SubscriptionInputBillingCycle];


export const SubscriptionInputBillingCycle = {
  weekly: 'weekly',
  monthly: 'monthly',
  quarterly: 'quarterly',
  semi_annual: 'semi_annual',
  yearly: 'yearly',
} as const;

export interface SubscriptionInput {
  /** @minLength 1 */
  name: string;
  logoUrl?: string;
  categoryId?: number;
  /** @minimum 0 */
  price: number;
  currency: string;
  billingCycle: SubscriptionInputBillingCycle;
  /** ISO date YYYY-MM-DD */
  renewalDate: string;
  paymentMethod?: string;
  notes?: string;
  isActive?: boolean;
}

export type SubscriptionUpdateBillingCycle = typeof SubscriptionUpdateBillingCycle[keyof typeof SubscriptionUpdateBillingCycle];


export const SubscriptionUpdateBillingCycle = {
  weekly: 'weekly',
  monthly: 'monthly',
  quarterly: 'quarterly',
  semi_annual: 'semi_annual',
  yearly: 'yearly',
} as const;

export interface SubscriptionUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @minimum 0 */
  price?: number;
  currency?: string;
  billingCycle?: SubscriptionUpdateBillingCycle;
  renewalDate?: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  notes?: string | null;
  isActive?: boolean;
}

export interface Category {
  id: number;
  name: string;
  color: string;
  icon: string;
  isDefault: boolean;
  /** @nullable */
  userId?: string | null;
  subscriptionCount?: number;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  color?: string;
  icon?: string;
}

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  color?: string;
  icon?: string;
}

export interface Reminder {
  id: number;
  userId: string;
  subscriptionId: number;
  /** @nullable */
  subscriptionName?: string | null;
  /** Days before renewal to send reminder (1, 3, 7) */
  daysBefore: number;
  isEnabled: boolean;
  createdAt: string;
}

export type ReminderInputDaysBefore = typeof ReminderInputDaysBefore[keyof typeof ReminderInputDaysBefore];


export const ReminderInputDaysBefore = {
  NUMBER_1: 1,
  NUMBER_3: 3,
  NUMBER_7: 7,
  NUMBER_14: 14,
  NUMBER_30: 30,
} as const;

export interface ReminderInput {
  subscriptionId: number;
  daysBefore: ReminderInputDaysBefore;
  isEnabled?: boolean;
}

export type ReminderUpdateDaysBefore = typeof ReminderUpdateDaysBefore[keyof typeof ReminderUpdateDaysBefore];


export const ReminderUpdateDaysBefore = {
  NUMBER_1: 1,
  NUMBER_3: 3,
  NUMBER_7: 7,
  NUMBER_14: 14,
  NUMBER_30: 30,
} as const;

export interface ReminderUpdate {
  daysBefore?: ReminderUpdateDaysBefore;
  isEnabled?: boolean;
}

export interface DashboardSummary {
  totalActiveSubscriptions: number;
  totalArchivedSubscriptions: number;
  monthlySpend: number;
  yearlySpend: number;
  /** Renewals in the next 7 days */
  upcomingRenewalsCount: number;
  /**
     * Days until next renewal, null if no subscriptions
     * @nullable
     */
  nextRenewalDays: number | null;
}

export interface CategorySpend {
  /** @nullable */
  categoryId: number | null;
  categoryName: string;
  color: string;
  icon: string;
  monthlyAmount: number;
  percentage: number;
  subscriptionCount: number;
}

export interface MonthlyTrend {
  month: number;
  year: number;
  /** e.g. "Jan 2025" */
  label: string;
  totalAmount: number;
  subscriptionCount: number;
}

/**
 * @nullable
 */
export type AnalyticsOverviewHighestExpense = {
  name?: string;
  monthlyEquivalent?: number;
  price?: number;
  billingCycle?: string;
} | null;

export interface AnalyticsOverview {
  totalAnnualSpend: number;
  averageMonthlySpend: number;
  /** @nullable */
  highestExpense: AnalyticsOverviewHighestExpense;
  averageSubscriptionCost: number;
  totalSubscriptions: number;
}

export interface CalendarDay {
  /** ISO date YYYY-MM-DD */
  date: string;
  subscriptions: Subscription[];
}

export type UserSettingsTheme = typeof UserSettingsTheme[keyof typeof UserSettingsTheme];


export const UserSettingsTheme = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export interface UserSettings {
  id: number;
  userId: string;
  /** @nullable */
  displayName?: string | null;
  currency: string;
  theme: UserSettingsTheme;
  timezone: string;
  createdAt: string;
  updatedAt: string;
}

export type UserSettingsUpdateTheme = typeof UserSettingsUpdateTheme[keyof typeof UserSettingsUpdateTheme];


export const UserSettingsUpdateTheme = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export interface UserSettingsUpdate {
  /** @nullable */
  displayName?: string | null;
  currency?: string;
  theme?: UserSettingsUpdateTheme;
  timezone?: string;
}

export type ListSubscriptionsParams = {
/**
 * Filter by status
 */
status?: ListSubscriptionsStatus;
/**
 * Filter by category ID
 * @nullable
 */
categoryId?: number | null;
/**
 * Filter by billing cycle
 */
billingCycle?: ListSubscriptionsBillingCycle;
/**
 * Search by name, notes
 */
search?: string;
/**
 * Minimum price filter
 * @nullable
 */
minPrice?: number | null;
/**
 * Maximum price filter
 * @nullable
 */
maxPrice?: number | null;
/**
 * Sort field
 */
sortBy?: ListSubscriptionsSortBy;
/**
 * Sort direction
 */
sortOrder?: ListSubscriptionsSortOrder;
};

export type ListSubscriptionsStatus = typeof ListSubscriptionsStatus[keyof typeof ListSubscriptionsStatus];


export const ListSubscriptionsStatus = {
  active: 'active',
  archived: 'archived',
  all: 'all',
} as const;

export type ListSubscriptionsBillingCycle = typeof ListSubscriptionsBillingCycle[keyof typeof ListSubscriptionsBillingCycle];


export const ListSubscriptionsBillingCycle = {
  weekly: 'weekly',
  monthly: 'monthly',
  quarterly: 'quarterly',
  semi_annual: 'semi_annual',
  yearly: 'yearly',
} as const;

export type ListSubscriptionsSortBy = typeof ListSubscriptionsSortBy[keyof typeof ListSubscriptionsSortBy];


export const ListSubscriptionsSortBy = {
  name: 'name',
  price: 'price',
  renewalDate: 'renewalDate',
  createdAt: 'createdAt',
  monthlyEquivalent: 'monthlyEquivalent',
} as const;

export type ListSubscriptionsSortOrder = typeof ListSubscriptionsSortOrder[keyof typeof ListSubscriptionsSortOrder];


export const ListSubscriptionsSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type GetCalendarEventsParams = {
year: number;
month: number;
};