# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SubTrack – Subscription Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories (defaults + user-created)
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a custom category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Validation error
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a custom category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a custom category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────────────────────────────
  # REMINDERS
  # ──────────────────────────────
  /reminders:
    get:
      operationId: listReminders
      tags: [reminders]
      summary: List all reminders for the current user
      responses:
        "200":
          description: List of reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"

    post:
      operationId: createReminder
      tags: [reminders]
      summary: Create a reminder for a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderInput"
      responses:
        "201":
          description: Created reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{id}:
    patch:
      operationId: updateReminder
      tags: [reminders]
      summary: Update a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderUpdate"
      responses:
        "200":
          description: Updated reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"

    delete:
      operationId: deleteReminder
      tags: [reminders]
      summary: Delete a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────────────────────────────
  # DASHBOARD
  # ──────────────────────────────
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, isDefault, color, icon]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        icon:
          type: string
        isDefault:
          type: boolean
        userId:
          type: ["string", "null"]
        subscriptionCount:
          type: integer
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
        icon:
          type: string
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
        icon:
          type: string

    # ── Reminder ──
    Reminder:
      type: object
      required: [id, userId, subscriptionId, daysBefore, isEnabled, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        subscriptionId:
          type: integer
        subscriptionName:
          type: ["string", "null"]
        daysBefore:
          type: integer
          description: Days before renewal to send reminder (1, 3, 7)
        isEnabled:
          type: boolean
        createdAt:
          type: string
    ReminderInput:
      type: object
      required: [subscriptionId, daysBefore]
      properties:
        subscriptionId:
          type: integer
        daysBefore:
          type: integer
          enum: [1, 3, 7, 14, 30]
        isEnabled:
          type: boolean
          default: true
    ReminderUpdate:
      type: object
      properties:
        daysBefore:
          type: integer
          enum: [1, 3, 7, 14, 30]
        isEnabled:
          type: boolean

    # ── Dashboard ──
...