# lib/api-zod/src/generated/types/reminder.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SubTrack – Subscription Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Reminder {
  id: number;
  userId: string;
  subscriptionId: number;
  /** @nullable */
  subscriptionName?: string | null;
  /** Days before renewal to send reminder (1, 3, 7) */
  daysBefore: number;
  isEnabled: boolean;
  createdAt: string;
}